+++ init.php
@@ -12,10 +12,11 @@ define( 'OPCM_PRODUCT_URL', 'https://github.com/Pierre-Lannoy/wp-opcache-manager
 define( 'OPCM_PRODUCT_SHORTNAME', 'OPcache' );
 define( 'OPCM_PRODUCT_ABBREVIATION', 'opcm' );
 define( 'OPCM_SLUG', 'opcache-manager' );
-define( 'OPCM_VERSION', '2.2.0' );
+define( 'OPCM_VERSION', '2.3.0-dev1' );
 define( 'OPCM_CODENAME', '"-"' );
 
 define( 'OPCM_CDN_AVAILABLE', true );
 
 define( 'OPCM_CRON_RESET_NAME', 'opcm_opcache_reset' );
+define( 'OPCM_CRON_WATCHDOG_NAME', 'opcm_watchdog' );
 define( 'OPCM_CRON_STATS_NAME', 'opcm_opcache_stats' );